Separate out the frontend structs from parser
