+++ components/support/nimbus-fml/src/backends/swift/mod.rs
@@ -3,7 +3,7 @@
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use crate::error::{FMLError, Result};
-use crate::parser::AboutBlock;
+use crate::frontend::AboutBlock;
 use askama::Template;
 
 use crate::commands::GenerateStructCmd;
