+++ components/support/nimbus-fml/src/commands.rs
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+use crate::frontend::AboutBlock;
 use crate::intermediate_representation::TargetLanguage;
-use crate::parser::AboutBlock;
 use crate::util::loaders::LoaderConfig;
 use anyhow::{bail, Error, Result};
 use std::path::Path;
